refactored items to new entities package
